requirements.txt
numpy
pandas
faiss-cpu
requests
langchain
streamlit
tf-keras
sentence_transformers
bs4
python-dotenv